fix: move regression tests to correct pipeline test file
